Separated and fix of the QR on the compiled version
