+++ options.cfg
@@ -24,7 +24,7 @@ USE_GESVD        = OFF
 USE_COMPILED     = ON
 # Comma separated list of the modules to be compiled
 # if USE_COMPILED = ON
-MODULES_COMPILED = MATH.MATHS,MATH.AVERAGING,MATH.SVD,MATH.FFT,MATH.GEOMETRIC,MATH.TRUNCATION,MATH.STATS,MATH.REGRESSION,ROM.POD,ROM.DMD,ROM.SPOD
+MODULES_COMPILED = MATH.MATHS,MATH.AVERAGING,MATH.QR,MATH.SVD,MATH.FFT,MATH.GEOMETRIC,MATH.TRUNCATION,MATH.STATS,MATH.REGRESSION,ROM.POD,ROM.DMD,ROM.SPOD
 
 
 ## Optimization, host and CPU type
